% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to.R
\name{bam_to}
\alias{bam_to}
\title{Convert a BAM file}
\usage{
bam_to(
  bam_files,
  formats = c("bedGraph", "bigWig"),
  width = 150,
  outdir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{formats}{A character vector of one or more formats for export file to.}

\item{width}{
    If \code{x} is an \link[IRanges]{IRanges}, \link[IRanges]{IPos} or \link[IRanges]{Views} object:
    For \code{narrow} and \code{threebands}, a vector of integers,
    eventually with NAs. See the SEW (Start/End/Width) interface for
    the details (\code{?solveUserSEW}).
    For \code{resize} and \code{flank}, the width of the resized or
    flanking regions. Note that if \code{both} is \code{TRUE}, this
    is effectively doubled. Recycled as necessary so that each
    element corresponds to a range in \code{x}.

    Can also be a list-like object \emph{parallel} to \code{x}
    if \code{x} is a \link[IRanges]{RangesList} object.
  }

\item{outdir}{Directory to save results to.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \code{rtracklayer::export}.}
}
\value{
Converted file paths.
}
\description{
Convert a BAM file to any format supported by \code{rtracklayer::export}.
}
\examples{
bam_files <- c(f1=system.file("extdata","querybins.bam",package="Rsamtools"),
               f2=system.file("extdata","revbins.bam",package="Rsamtools"))
converted_files <- bam_to(bam_files = bam_files)               
}
