% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks_macsr.R
\name{call_peaks_macsr}
\alias{call_peaks_macsr}
\title{Call peaks: MACSr}
\usage{
call_peaks_macsr(
  bedgraph_path,
  cutoff = NULL,
  minlen = 200L,
  maxgap = 30L,
  call_summits = TRUE,
  trackline = TRUE,
  log = TRUE,
  outdir = tempdir(),
  outputfile = "MACSr.peaks.bed",
  return_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bedgraph_path}{Path to bedGraph file. 
Can instead provide a bigWig file, 
but this will first be converted to bedGraph format, 
which can take some time if trying to convert data from across the entire
genome.}

\item{cutoff}{Cutoff depends on which method you used for score track.
If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
through a \code{cutoff_analysis}.}

\item{minlen}{minimum length of peak, better to set it as d
value. DEFAULT: 200", default = 200.}

\item{maxgap}{maximum gap between significant points in a peak,
better to set it as tag size. DEFAULT: 30", default = 30.}

\item{call_summits}{If set, MACS will use a more sophisticated
approach to find all summits in each enriched peak region
DEFAULT: False",default=False.}

\item{trackline}{Tells MACS not to include trackline with bedGraph
files. The trackline is required by UCSC.}

\item{log}{Whether to capture logs.}

\item{outdir}{Directory to store \code{cutoff_analysis} 
report and peak file in.}

\item{outputfile}{Name of the peak output file (stored in BED format).}

\item{return_path}{Whether to return the path to the saved peak file, 
or the peak data itself as a \link[GenomicRanges]{GRanges} object.}

\item{verbose}{Print messages.}
}
\value{
\link[GenomicRanges]{GRanges} or path to save peaks file.
}
\description{
Call peaks from a bedGraph or bigWig file using multiple methods.
By default, it automatically infers a reasonable 
\code{cutoff} threshold as well.\cr
\emph{Note : } 
\pkg{MACS3}/\pkg{MACSr} is not currently compatible with Windows 
(see \href{https://github.com/macs3-project/MACSr/issues/13}{here}
 for details).
}
\keyword{internal}
