% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_peaks_seacr.R
\name{call_peaks_seacr}
\alias{call_peaks_seacr}
\title{Call peaks: SEACR}
\source{
\href{https://github.com/FredHutch/SEACR/issues/54}{
GitHub Issue: illegal byte sequence}
}
\usage{
call_peaks_seacr(
  bedgraph_path,
  control = 0.05,
  norm = TRUE,
  stringent = TRUE,
  outdir = tempdir(),
  outputfile = basename(tempfile(pattern = bedgraph_path, fileext = ".SEACR_peaks")),
  return_path = FALSE,
  remove_tmps = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{bedgraph_path}{Field 1: Target data bedgraph file in 
\href{https://genome.ucsc.edu/goldenpath/help/bedgraph.html}{
UCSC bedgraph format} that omits regions containing 0 signal.}

\item{control}{Control (IgG) data bedgraph file to generate an empirical
threshold for peak calling.
Alternatively, a numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks 
(default: \code{0.05}).}

\item{norm}{Field 3: "norm" (default: \code{norm=TRUE}) 
denotes normalization of control 
to target data, "non" (\code{norm=FALSE}) skips this behavior.
"norm" is recommended unless experimental and control data are already 
rigorously normalized to each other (e.g. via spike-in).}

\item{stringent}{Field 4:"relaxed" (\code{stringent=FALSE})
uses a total signal 
threshold between the knee and peak of the total signal curve,
and corresponds to the “relaxed” mode described in the text, 
whereas “stringent” (default: \code{stringent=TRUE}) 
uses the peak of the curve, and corresponds to “stringent” mode.}

\item{outdir}{Directory to store \code{cutoff_analysis} 
report and peak file in.}

\item{outputfile}{Name of the peak output file (stored in BED format).}

\item{return_path}{Whether to return the path to the saved peak file, 
or the peak data itself as a \link[GenomicRanges]{GRanges} object.}

\item{remove_tmps}{Remove temporary intermediate files generated by 
\pkg{SEACR} (i.e. )}

\item{nThread}{When \code{nThread>1}, accelerates file importing 
and peak calling using multi-core parallelisation.}

\item{verbose}{Print messages.}
}
\value{
\link[GenomicRanges]{GRanges} or path to save peaks file.
}
\description{
Call peaks from a bedGraph or bigWig file using
\href{https://github.com/FredHutch/SEACR}{SEACR}
(Sparse Enrichment Analysis for CUT&RUN).
}
\keyword{internal}
