% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_search.R
\name{construct_searches}
\alias{construct_searches}
\title{Construct searches}
\usage{
construct_searches(searches = NULL, keys = NULL)
}
\arguments{
\item{searches}{Named list of regex queries.}

\item{keys}{Keys to subset the \code{searches} with by their
names.}
}
\description{
Construct a named list to be used to identify matching file types
based on their path names.
}
\examples{
searches <- construct_searches()
}
