% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bam.R
\name{convert_bam}
\alias{convert_bam}
\title{BAM to bedGraph}
\usage{
convert_bam(
  bam_files,
  formats = c("bedGraph", "bigWig"),
  width = 150,
  verbose = TRUE
)
}
\arguments{
\item{bam_files}{One or more BAM files.}
}
\value{
Converted file paths.
}
\description{
Convert a BAM file to any format supported by \link[rtracklayer]{export}.
}
\examples{
bam_files <- c(f1=system.file("extdata","querybins.bam",package="Rsamtools"),
               f2=system.file("extdata","revbins.bam",package="Rsamtools"))
converted_files <- convert_bam(bam_files = bam_files)               
}
