% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trackhub.R
\name{create_trackhub}
\alias{create_trackhub}
\title{Create TrackHub}
\source{
\href{UCSC track metadata descriptions}{
https://genome-euro.ucsc.edu/goldenPath/help/trackDb/trackDbHub.html
}

\href{example trackhub}{
https://github.com/mhalushka/UCSC_miRNA_barchart
}

\href{https://github.com/lawremi/rtracklayer/issues/12}{
rtracklayer possible memory leakage
}
}
\usage{
create_trackhub(
  hub = "my_hub_name",
  genome,
  path = getwd(),
  save_dir = file.path(path, "trackhub"),
  folders = c(""),
  domain = paste0("https://neurogenomics-ukdri.dsi.ic.ac.uk/", hub),
  shortLabel = "",
  longLabel = "",
  email = "",
  descriptionUrl = paste(save_dir, "description.html", sep = "/"),
  filetypes = c(bed = ".bed$", bigWig = ".bigwig$|.bw$"),
  settings = list(visibility = 2, fontSize = "10", height = "100", autoScale = "on",
    color = "255,102,102", renderer = "BAR_CHART", smoothingWindow = "off",
    windowFunction = "mean", logo = "on"),
  as_list = TRUE,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{genome}{Genome build that files were aligned to.}

\item{path}{Path to search for files within recursively.}

\item{folders}{Subfolders within \code{path} to search,}

\item{domain}{URL prefix where the files will be hosted.}

\item{filetypes}{File types to search for.}

\item{settings}{Trackhub settings. 
See 
\href{ https://genome-euro.ucsc.edu/goldenPath/help/trackDb/trackDbHub.html}{
UCSC trackhub documentation} for details.}

\item{as_list}{Return each file as a nested list.}

\item{save_path}{Path to to save \emph{trackDb} file to.}
}
\value{
Character or nested list.
}
\description{
Create TrackHub for visualization of genomic files 
in the UCSC Genome Browser.
}
\examples{
\dontrun{
path <-"/Volumes/bms20/projects/neurogenomics-lab/live/Projects/CUT_n_TAG/CUTnTag_analysis"
res <- create_trackhub(path=path, folders = c("bigwig","bedgraph","peaks"))

path <-"/Volumes/bms20/projects/neurogenomics-lab/live/Data/tip_seq/processed_data/without_duplicates/tipseq_9_17_jun_2022/"
res <- create_trackhub(hub="tipseq_9_17_jun_2022",
                       path=path, 
                       folders = "03_peak_calling",
                       genome="hg38")
}
}
