% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_bam.R
\name{example_bam}
\alias{example_bam}
\title{Example BAM files}
\usage{
example_bam(save_dir = tempdir())
}
\arguments{
\item{save_dir}{Directory to save files to.}
}
\description{
Download exampe BAM files from 
\href{nf-core}{https://nf-co.re/cutandrun/results}.
}
\examples{
bam_files <- example_bam()
}
