% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_executable_seacr.R
\name{find_executable_seacr}
\alias{find_executable_seacr}
\title{Find executable: \pkg{SEACR}}
\source{
\href{https://github.com/FredHutch/SEACR/issues/54}{
GitHub Issue: illegal byte sequence}
}
\usage{
find_executable_seacr(
  packages = c("SEACR_1.3.sh", "SEACR_1.3.R"),
  fix_script = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{packages}{A package (e.g. \code{"pandas"})
or a list of packages (e.g. \code{c("pandas","r-dplyr","bgzip")}).
Will recognize R packages with or without the prefixes 
"r-" and "bioconductor-" (i.e. "r-dplyr" is equivalent to "dplyr").}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[echoconda:find_packages]{echoconda::find_packages}}
  \describe{
    \item{\code{conda_env}}{Conda environments to search in. 
If \code{NULL} (default), will search all conda environments.}
    \item{\code{conda}}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}
    \item{\code{method}}{Method to use:
\itemize{
\item{"basilisk"}
\item{"reticulate"}
}}
    \item{\code{types}}{Path type to search for and include in the table.}
    \item{\code{return_path}}{Return only the path to each package.}
    \item{\code{robust}}{Uses a more robust (but slower) method for identifying paths
to conda environments and package executables (default: \code{FALSE}).
When \code{robust=FALSE}, only executables in the main "bin" directory
 will be detected. When \code{robust=TRUE}, additional executables to
 R packages, Python packages, and executables in nested subdirectories will 
 be detected.}
    \item{\code{filter_paths}}{Filter out packages without callable executable paths
from the table.}
    \item{\code{sort_names}}{Sort packages alphanumerically by their name.}
    \item{\code{nThread}}{Number of threads to use when parallelizing searches across
multiple conda envs.}
  }}
}
\value{
Path to \pkg{SEACR} executable.
}
\description{
Find path to the executable file for \pkg{SEACR}.
}
\keyword{internal}
