% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_links.R
\name{find_links}
\alias{find_links}
\title{Find links}
\usage{
find_links(urls, pattern = NULL, as_datatable = TRUE, workers = 1)
}
\arguments{
\item{urls}{List of URLs to search for links on.}

\item{pattern}{Regex pattern to search for files with.}

\item{as_datatable}{Return the results as a \link[data.table]{data.table}
(default: \code{TRUE}), or a nested list (\code{FALSE}).}

\item{workers}{Number of threads of parallelize across.}
}
\value{
A data.table or named list.
}
\description{
Scrape a web page for links.
}
\examples{
links <- find_links(
    urls="http://renlab.sdsc.edu/kai/Key_Processed_Data/ABC_scores/",
    pattern="\\\\.tsv\\\\.gz")
}
