% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encode_peaks_metadata.R
\name{get_encode_peaks_metadata}
\alias{get_encode_peaks_metadata}
\title{Get ENCODE peaks metadata}
\source{
\code{
### Get peak data only #####
meta <- search_encode(file_type = "peak",
                      assembly = "GRCh38",
                      organism = "Homo sapiens",
                      partial_match = TRUE)
for(col in names(meta)){
    if(all(is.na(meta[[col]]))){
        messager("Removing:",col)
        meta[[col]] <- NULL
    }
}
tmp <- "~/Downloads/encode_peaks_metadata.tsv.gz"
data.table::fwrite(meta, file = tmp, sep="\t")
piggyback::pb_new_release(repo = "neurogenomics/PeakyFinders",
                          tag="latest")
piggyback::pb_upload(file = tmp,
                     repo = "neurogenomics/PeakyFinders",
                     overwrite = TRUE)
}
}
\usage{
get_encode_peaks_metadata(
  save_dir = tools::R_user_dir(package = "PeakyFinders", which = "cache"),
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
}
\description{
Get a subset of the ENCODE metadata for epigenomic peaks in human data.
}
\keyword{internal}
