% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome.R
\name{get_genome}
\alias{get_genome}
\title{Get genome}
\usage{
get_genome(
  keep.chr = paste0("chr", c(seq_len(22), "X", "Y")),
  genome = "hg19",
  sort = TRUE,
  style = "UCSC",
  split_chromosomes = FALSE
)
}
\arguments{
\item{keep.chr}{is a character vector stating the names of the chromosomes to keep. Any chromosome not in the vector will be filtered out. If keep.chr is supplied, organism and chr.type are ignored.}

\item{genome}{The genome object or genome identifier.}

\item{sort}{Whether to sort the \link[GenomicRanges]{GRanges}
by chromosomes.}

\item{style}{Chromosome style,
set by \link[GenomeInfoDb]{seqlevelsStyle}.
\itemize{
\item{"UCSC" : }{Uses the chromosome style "chr1".}
\item{"NCBI" : }{Uses the chromosome style "1"}
}}

\item{split_chromosomes}{Whether to split the 
\link[GenomicRanges]{GRanges} by chromosome (as a named list).}
}
\value{
\link[GenomicRanges]{GRanges}
}
\description{
Get a genome (or a subset of chromosomes from a genome) 
as a \link[GenomicRanges]{GRanges} object. This can be useful for 
querying whole-chromosomes at a time.
}
\examples{
gr <- get_genome(keep.chr=1:2)
}
