% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_links.R
\name{get_geo_links}
\alias{get_geo_links}
\title{Get GEO supplementary files.}
\usage{
get_geo_links(gsm, searches = construct_searches(), verbose = TRUE)
}
\arguments{
\item{gsm}{GEO GSM id.
Matches will be case-insensitive.}

\item{searches}{Named list of regex queries.}

\item{verbose}{Print messages.}
}
\value{
Named list.
}
\description{
Get links to any GEO supplementary files matches one or more of the 
\code{searches}.
\emph{IMPORTANT:} When downloading files from GEO, make sure
"ftp://" are replaced with "http://" in URLs, as the former can
cause download issues on certain machines.
}
\keyword{internal}
