% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_supplementary_files.R
\name{get_geo_supplementary_files}
\alias{get_geo_supplementary_files}
\title{Get GEO supplementary files.}
\usage{
get_geo_supplementary_files(
  gsm,
  searches = construct_searches(),
  verbose = TRUE
)
}
\arguments{
\item{gsm}{GEO GSM id.
Matches will be case-insensitive.}

\item{searches}{Named list of regex queries.}

\item{verbose}{Print messages.}
}
\description{
Get links to any GEO supplementary files matches one or more of the 
\code{searches}.
}
\keyword{internal}
