% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaks_annotationhub.R
\name{import_peaks_annotationhub}
\alias{import_peaks_annotationhub}
\title{Import peaks: AnnotationHub}
\usage{
import_peaks_annotationhub(
  ids,
  build = "hg19",
  query_granges = NULL,
  query_granges_build = NULL,
  split_chromosomes = FALSE,
  condense_queries = TRUE,
  force_new = FALSE,
  call_peaks_method = "MACSr",
  cutoff = NULL,
  searches = construct_searches(),
  peaks_dir = tempdir(),
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{IDs from one of the supported databases. 
IDs can be at any level: file, sample, or experiment.}

\item{query_granges}{[Optional]
\link[GenomicRanges]{GRanges} object indicating which genomic
regions to extract from each sample.}

\item{query_granges_build}{[Optional]
Genome build that \code{query_granges} is aligned to.}

\item{split_chromosomes}{Split single-threaded query 
into multi-threaded query across chromosomes. 
This is can be helpful especially when calling peaks from 
large bigWig/bedGraph files.
The number of threads used is set by the \code{nThread} argument.}

\item{condense_queries}{Condense \code{query_granges}
by taking the min/max position per chromosome (default: \code{TRUE)}.
This helps to reduce the total number of queries, 
which can cause memory allocation problems 
due to repeated calls to the underlying C libraries.}

\item{force_new}{By default, saved results of the same \code{save_path} name 
will be imported instead of running queries. However you can override this
by setting \code{force_new} to perform new queries regardless and overwrite
 the old \code{save_path} file.}

\item{call_peaks_method}{Method to call peaks with:
\itemize{
\item{"MACSr" : }{Uses \href{https://github.com/macs3-project/MACS}{MACS3} 
via \link[MACSr]{bdgpeakcall}.}
}}

\item{cutoff}{\itemize{
\item{when \code{call_peaks_method="MACSr"} : }{
Passed to \code{cutoff} argument.
Cutoff depends on which method you used for score track.
 If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
 If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
 through a \code{cutoff_analysis}. 
}
\item{when \code{call_peaks_method="SEACR"} : }{
Passed to \code{control} argument.
Control (IgG) data bedgraph file to generate an empirical
threshold for peak calling.
Alternatively, a numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks 
(default: \code{0.05}).
}
}}

\item{searches}{Named list of regex queries.}

\item{peaks_dir}{Directory to save peaks to 
(only used when calling peaks from bedGraph files).}

\item{nThread}{When \code{nThread>1}, accelerates file importing 
and peak calling using multi-core parallelisation.}

\item{verbose}{Print messages.}
}
\value{
\link[GenomicRanges]{GRanges}
}
\description{
Import peaks from \href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{
AnnotationHub}.
}
\keyword{internal}
