% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaks_geo.R
\name{import_peaks_geo}
\alias{import_peaks_geo}
\title{Import peaks: GEO}
\source{
\href{https://github.com/ropensci/rtweet/issues/229}{
curl connection timeout issues}

\href{https://github.com/Bioconductor/BiocParallel/issues/111}{
Issue with importing bigWig files in parallel}
}
\usage{
import_peaks_geo(
  ids,
  build,
  query_granges,
  query_granges_build,
  split_chromosomes = FALSE,
  method = "MACSr",
  cutoff = NULL,
  searches = construct_searches(),
  condense_queries = TRUE,
  peaks_dir = tempdir(),
  timeout = 3 * 60,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{ids}{IDs from one of the supported databases. 
IDs can be at any level: file, sample, or experiment.}

\item{query_granges}{[Optional]
\link[GenomicRanges]{GRanges} object indicating which genomic
regions to extract from each sample.}

\item{query_granges_build}{[Optional]
Genome build that \code{query_granges} is aligned to.}

\item{split_chromosomes}{Split single-threaded query 
into multi-threaded query across chromosomes. 
This is can be helpful especially when calling peaks from 
large bigWig/bedGraph files.
The number of threads used is set by the \code{nThread} argument.}

\item{method}{Method to call peaks with:
\itemize{
\item{"MACSr" : }{Uses \href{https://github.com/macs3-project/MACS}{MACS3} 
via \link[MACSr]{bdgpeakcall}.}
\item{"SEACR" : }{Uses \href{https://github.com/FredHutch/SEACR}{SEACR} 
via \link[echoconda]{find_packages}.}
}}

\item{cutoff}{\itemize{
\item{when \code{method="MACSr"} : }{
Passed to \code{cutoff} argument.
Cutoff depends on which method you used for score track.
 If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
 If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
 through a \code{cutoff_analysis}. 
}
\item{when \code{method="SEACR"} : }{
Passed to \code{control} argument.
Control (IgG) data bedgraph file to generate an empirical
threshold for peak calling.
Alternatively, a numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks 
(default: \code{0.05}).
}
}}

\item{searches}{Named list of regex queries.}

\item{condense_queries}{Condense \code{query_granges}
by taking the min/max position per chromosome (default: \code{TRUE)}.
This helps to reduce the total number of queries, 
which can cause memory allocation problems 
due to repeated calls to the underlying C libraries.}

\item{peaks_dir}{Directory to save peaks to 
(only used when calling peaks from bedGraph files).}

\item{nThread}{When \code{nThread>1}, accelerates file importing 
and peak calling using multi-core parallelisation.}

\item{verbose}{Print messages.}

\item{gsm}{GEO GSM id (e.g. "GSM4271282").}
}
\value{
Named list of peak files in \link[GenomicRanges]{GRanges} format.
}
\description{
Import narrow/broad/generic peaks from GEO, or compute peaks with 
\link[PeakyFinders]{call_peaks}.
}
\details{
Must import \link[methods]{new} in my function
because it seems \link[GEOquery]{getGEO} forgot to do this 
(only works when you load the entire \pkg{GEOquery} package first).
}
\keyword{internal}
