% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaks_multi.R
\name{import_peaks_multi}
\alias{import_peaks_multi}
\title{Import peaks: multi}
\usage{
import_peaks_multi(
  links,
  id = basename(tempfile()),
  build = "hg19",
  query_granges = NULL,
  query_granges_build = NULL,
  split_chromosomes = FALSE,
  condense_queries = TRUE,
  call_peaks_method = "MACSr",
  cutoff = NULL,
  searches = construct_searches(),
  peaks_dir = tempdir(),
  merge_list = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{query_granges}{[Optional]
\link[GenomicRanges]{GRanges} object indicating which genomic
regions to extract from each sample.}

\item{query_granges_build}{[Optional]
Genome build that \code{query_granges} is aligned to.}

\item{split_chromosomes}{Split single-threaded query 
into multi-threaded query across chromosomes. 
This is can be helpful especially when calling peaks from 
large bigWig/bedGraph files.
The number of threads used is set by the \code{nThread} argument.}

\item{condense_queries}{Condense \code{query_granges}
by taking the min/max position per chromosome (default: \code{TRUE)}.
This helps to reduce the total number of queries, 
which can cause memory allocation problems 
due to repeated calls to the underlying C libraries.}

\item{call_peaks_method}{Method to call peaks with:
\itemize{
\item{"MACSr" : }{Uses \href{https://github.com/macs3-project/MACS}{MACS3} 
via \link[MACSr]{bdgpeakcall}.}
}}

\item{cutoff}{Cutoff depends on which method you used for score track.
If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
through a \code{cutoff_analysis}.}

\item{searches}{Named list of regex queries.}

\item{peaks_dir}{Directory to save peaks to 
(only used when calling peaks from bedGraph files).}

\item{nThread}{When \code{nThread>1}, accelerates file importing 
and peak calling using multi-core parallelisation.}

\item{verbose}{Print messages.}
}
\value{
\link[GenomicRanges]{GRanges}
}
\description{
Automatically import peaks in different file formats.
If the file is in \emph{bigWig} or \emph{bedGraph} format, 
peaks wil be called first.
}
\keyword{internal}
