% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_local.R
\name{is_local}
\alias{is_local}
\title{Is local}
\usage{
is_local(path)
}
\arguments{
\item{path}{Path to local file or remote URL.}
}
\description{
Determine whether a file exist locally.
}
\keyword{internal}
