% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_url}
\alias{is_url}
\title{Check if the input is url e.g. http:// or ftp://}
\source{
\href{https://rdrr.io/cran/seqminer/src/R/seqminer.R}{
Borrowed from \code{seqminer} internal function}
}
\usage{
is_url(fileName)
}
\arguments{
\item{fileName}{character vector}
}
\description{
Check if the input is url e.g. http:// or ftp://
}
\keyword{internal}
