% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conda_env.R
\name{make_conda_env}
\alias{make_conda_env}
\title{Make conda environment}
\usage{
make_conda_env(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[echoconda:yaml_to_env]{echoconda::yaml_to_env}}
  \describe{
    \item{\code{yaml_path}}{Path to local or remote yaml file with conda build
specifications.}
    \item{\code{method}}{Method to use to create the conda env:
\itemize{
\item{"basilisk" : }{Uses the R package \pkg{basilisk}.}
\item{"reticulate" : }{Uses the R package \pkg{reticulate}.}
\item{"cli" : }{Uses a custom wrapper for conda's command line interface.}
}}
    \item{\code{conda}}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}
    \item{\code{force_new}}{If the conda env already exists,
overwrite it with a new one (\emph{DEFAULT}: \code{FALSE}).}
    \item{\code{show_contents}}{Show the contents of the yaml file (if used).}
    \item{\code{verbose}}{Print messages.}
  }}
}
\value{
Conda env name.
}
\description{
Make a dedicated conda environment for \pkg{PeakyFinders}.
}
\keyword{internal}
