% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_bam.R
\name{merge_bam}
\alias{merge_bam}
\title{Merge BAM files}
\usage{
merge_bam(
  bam_files,
  destination = tempfile(fileext = ".merged.bam"),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{destination}{The character(1) file name of the location where
    the sorted, filtered, or merged output file will be created. For
    \code{asBam} \code{asSam}, and \code{sortBam} this is without the
    \dQuote{.bam} file suffix.}

\item{overwrite}{A logical(1) indicating whether the destination
    can be over-written if it already exists.}

\item{verbose}{Print messages.}
}
\value{
Path to merged BAM file.
}
\description{
If >1 BAM file provided, merges them into a single BAM file using 
\link[Rsamtools]{mergeBam}.
}
\examples{
bam_files <- c(system.file("extdata", "querybins.bam", package="Rsamtools"),
               system.file("extdata", "revbins.bam", package="Rsamtools"))
bam_merged <- merge_bam(bam_files = bam_files)
}
