% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks_metadata_annotationhub}
\alias{peaks_metadata_annotationhub}
\title{Metadata for epigenomic peaks: AnnotationHub}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12483 rows and 24 columns.
}
\source{
\code{
meta <- search_annotationhub()
searches <- construct_searches(keys = c("broadpeak","narrowpeak","genericpeak"))
search_strings <- paste(unname(unlist(searches)),collapse = "|")
peaks_metadata_annotationhub <- meta[with(meta,
                  grepl(search_strings,
                        paste(title,description,sourcetype,sourceurl),
                        ignore.case = TRUE)
),
]
PeakyFinders:::drop_empty_cols(peaks_metadata_annotationhub)
usethis::use_data(peaks_metadata_annotationhub, overwrite = TRUE)
}
}
\usage{
data("peaks_metadata_annotationhub")
}
\description{
A subset of metadata from AnnotationHub that only includes 
 entries for peaks (narrowPeak and broadPeak).
}
\keyword{datasets}
