% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks_metadata_roadmap}
\alias{peaks_metadata_roadmap}
\title{Metadata for epigenomic peaks: ROADMAP}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3146 rows and 23 columns.
}
\source{
\code{
meta <- search_roadmap()
peaks_metadata_roadmap <- meta[output_category=="peaks" & 
                               (!is.na(biosample_id)),]
usethis::use_data(peaks_metadata_roadmap, overwrite = TRUE)
}
}
\usage{
data("peaks_metadata_roadmap")
}
\description{
A subset of metadata from the Encyclopedia of DNA Elements
 (\href{https://www.encodeproject.org/}{ENCODE}) that only includes 
 entries for peaks (narrowPeak and broadPeak). 
 Metadata originally sourced from \link[AnnotationHub]{AnnotationHub} with
 additional post-processing and filtering.
}
\keyword{datasets}
