% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_peaks.R
\name{pooled_peaks}
\alias{pooled_peaks}
\title{Pooled peaks}
\usage{
pooled_peaks(
  bam_files,
  groups = names(bam_files),
  outdir = tempdir(),
  cutoff = NULL,
  method = c("MACSr", "SEACR"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{groups}{A character vector of the same length as \code{bam_files}
defining how to group files when calling consensus peaks.}

\item{outdir}{Directory to save results to.}

\item{cutoff}{\itemize{
\item{when \code{method="MACSr"} : }{
Passed to \code{cutoff} argument.
Cutoff depends on which method you used for score track.
 If the file contains pvalue scores from MACS3, score 5 means pvalue 1e-5.
 If \code{NULL}, a reasonable \code{cutoff} value will be inferred 
 through a \code{cutoff_analysis}. 
}
\item{when \code{method="SEACR"} : }{
Passed to \code{control} argument.
Control (IgG) data bedgraph file to generate an empirical
threshold for peak calling.
Alternatively, a numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks 
(default: \code{0.05}).
}
}}

\item{method}{Method to call consensus peaks with:
\itemize{
\item{"MACSr : }{Call consensus peaks by merging groups of BAM files with 
\link[Rsamtools]{mergeBam} and then calling peaks with 
\link[PeakyFinders]{call_peaks_macsr}.
}
\item{"SEACR : }{Call consensus peaks by merging groups of BAM files with 
\link[Rsamtools]{mergeBam} and then calling peaks with 
\link[PeakyFinders]{call_peaks_seacr}.
}
}}

\item{verbose}{Print messages.}

\item{...}{Additional argument passed to either 
\link[PeakyFinders]{call_peaks_macsr} or
\link[PeakyFinders]{call_peaks_seacr}, depending on \code{method}.}
}
\value{
\link[GenomicRanges]{GRangesList} object.
}
\description{
Pool groups of BAM files into a smaller number of BAM files and call peaks
on each of them.
}
\examples{
bam_files <- example_bam()
peaks <- pooled_peaks(bam_files = bam_files)
}
