% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_peaks_seacr.R
\name{pooled_peaks_seacr}
\alias{pooled_peaks_seacr}
\title{Call consensus peaks: \pkg{SEACR}}
\usage{
pooled_peaks_seacr(
  bam_files,
  outdir = tempdir(),
  g = NULL,
  overwrite = TRUE,
  nThread = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{outdir}{Directory to save results to.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[=call_peaks_seacr]{call_peaks_seacr}}
  \describe{
    \item{\code{bedgraph_path}}{Field 1: Target data bedgraph file in 
\href{https://genome.ucsc.edu/goldenpath/help/bedgraph.html}{
UCSC bedgraph format} that omits regions containing 0 signal.}
    \item{\code{control}}{Control (IgG) data bedgraph file to generate an empirical
threshold for peak calling.
Alternatively, a numeric threshold n between 0 and 1 returns the top n
fraction of peaks based on total signal within peaks 
(default: \code{0.05}).}
    \item{\code{norm}}{Field 3: "norm" (default: \code{norm=TRUE}) 
denotes normalization of control 
to target data, "non" (\code{norm=FALSE}) skips this behavior.
"norm" is recommended unless experimental and control data are already 
rigorously normalized to each other (e.g. via spike-in).}
    \item{\code{stringent}}{Field 4:"relaxed" (\code{stringent=FALSE})
uses a total signal 
threshold between the knee and peak of the total signal curve,
and corresponds to the “relaxed” mode described in the text, 
whereas “stringent” (default: \code{stringent=TRUE}) 
uses the peak of the curve, and corresponds to “stringent” mode.}
    \item{\code{remove_tmps}}{Remove temporary intermediate files generated by 
\pkg{SEACR} (i.e. )}
    \item{\code{nThread}}{When \code{nThread>1}, accelerates file importing 
and peak calling using multi-core parallelisation.}
    \item{\code{outputfile}}{Name of the peak output file (stored in BED format).}
    \item{\code{return_path}}{Whether to return the path to the saved peak file, 
or the peak data itself as a \link[GenomicRanges]{GRanges} object.}
  }}
}
\value{
\link[GenomicRanges]{GRanges} object.
}
\description{
Call consensus peaks by merging groups of BAM files with 
\link[Rsamtools]{mergeBam} and then calling peaks with 
\link[PeakyFinders]{call_peaks_seacr}.
}
\keyword{internal}
