% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_annotationhub.R
\name{search_annotationhub}
\alias{search_annotationhub}
\title{Search metadata: AnnotationHub}
\source{
{https://biodatascience.github.io/compbio/bioc/anno.html}{
Mike Love tutorial on accessing peak data from AnnotationHub.}
}
\usage{
search_annotationhub(
  searches = construct_searches(),
  ah_id = NULL,
  output_category = NULL,
  output_type = NULL,
  file_format_type = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  dataset_biosample_summary = NULL,
  dataprovider = NULL,
  species = NULL,
  genome = NULL,
  description = NULL,
  rdataclass = NULL,
  sourcetype = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{searches}{Named list of regex queries.}

\item{ah_id}{Search "ah_id" column.}

\item{output_category}{Search "output_category" column.}

\item{output_type}{Search "output_type" column.}

\item{file_format_type}{Search "file_format_type" column.}

\item{biosample_id}{Search "biosample_id" column.}

\item{biosample_type}{Search "biosample_type" column.}

\item{biosample_name}{Search "biosample_name" column.}

\item{dataset_biosample_summary}{Search "dataset_biosample_summary" column.}

\item{dataprovider}{Search "dataprovider" column.}

\item{species}{Search "species" column.}

\item{genome}{Search "genome" column.}

\item{description}{Search "description" column.}

\item{rdataclass}{Search "rdataclass" column.}

\item{sourcetype}{Search "sourcetype" column.}

\item{partial_match}{Return case-insensitive substring matches instead of 
exact matches only (default: \code{TRUE}).}

\item{peaks_only}{Return pre-filtered metadata for peak data 
aligned to the GRCh38 human genome build.}

\item{verbose}{Print messages.}

\item{...}{Additional metadata columns to filter.}
}
\value{
\link[data.table]{data.table}.
}
\description{
Search metadata from 
\href{https://doi.org/doi:10.18129/B9.bioc.AnnotationHub}{AnnotationHub}.
}
\examples{
meta <- search_annotationhub(searches=construct_searches(keys="narrowpeak"),
                             biosample_type="brain",
                             peaks_only=TRUE)
}
