% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_encode.R
\name{search_encode}
\alias{search_encode}
\title{Search metadata: ENCODE}
\source{
{https://github.com/neurogenomics/EpiCompare/issues/43}{
Discussion on EpiCompare GitHub.}

{https://biodatascience.github.io/compbio/bioc/anno.html}{
Mike Love tutorial on accessing peak data from AnnotationHub.}
}
\usage{
search_encode(
  assay = NULL,
  accession = NULL,
  file_accession = NULL,
  file_type = NULL,
  file_format = NULL,
  output_category = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  organism = NULL,
  dataset_type = NULL,
  assembly = NULL,
  genome_annotation = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{assay}{Search "assay" column.}

\item{accession}{Search "accession" column.}

\item{file_accession}{Search "file_accession" column.}

\item{file_type}{Search "file_type" column.}

\item{file_format}{Search "file_format" column.}

\item{output_category}{Search "output_category" column.}

\item{biosample_id}{Search "biosample_id" column.}

\item{biosample_type}{Search "biosample_type" column.}

\item{biosample_name}{Search "biosample_name" column.}

\item{organism}{Search "organism" column.}

\item{dataset_type}{Search "dataset_type" column.}

\item{assembly}{Search "assembly" column.}

\item{genome_annotation}{Search "genome_annotation" column.}

\item{partial_match}{Return case-insensitive substring matches instead of 
exact matches only (default: \code{TRUE}).}

\item{peaks_only}{Return pre-filtered metadata for peak data 
aligned to the GRCh38 human genome build.}

\item{verbose}{Print messages.}

\item{...}{Additional metadata columns to filter.}
}
\value{
\link[data.table]{data.table}.
}
\description{
Search metadata from the Encyclopedia of DNA Elements
 (\href{https://www.encodeproject.org/}{ENCODE}).
}
\examples{
meta <- search_encode(assay = "chip-seq", 
                      biosample_name = "K562",
                      file_type = "peak",
                      assembly = "GRCh38",
                      organism = "Homo sapiens",
                      partial_match = TRUE)
}
