% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_roadmap.R
\name{search_roadmap}
\alias{search_roadmap}
\title{Search metadata: ENCODE}
\source{
{https://github.com/neurogenomics/EpiCompare/issues/43}{
Discussion on EpiCompare GitHub.}

{https://biodatascience.github.io/compbio/bioc/anno.html}{
Mike Love tutorial on accessing peak data from AnnotationHub.}
}
\usage{
search_roadmap(
  output_category = NULL,
  output_type = NULL,
  file_format_type = NULL,
  biosample_id = NULL,
  biosample_type = NULL,
  biosample_name = NULL,
  dataset_biosample_summary = NULL,
  dataprovider = NULL,
  species = NULL,
  genome = NULL,
  description = NULL,
  rdataclass = NULL,
  sourcetype = NULL,
  partial_match = TRUE,
  peaks_only = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{output_category}{Search "output_category" column.}

\item{output_type}{Search "output_type" column.}

\item{file_format_type}{Search "file_format_type" column.}

\item{biosample_id}{Search "biosample_id" column.}

\item{biosample_type}{Search "biosample_type" column.}

\item{biosample_name}{Search "biosample_name" column.}

\item{dataset_biosample_summary}{Search "dataset_biosample_summary" column.}

\item{dataprovider}{Search "dataprovider" column.}

\item{species}{Search "species" column.}

\item{genome}{Search "genome" column.}

\item{description}{Search "description" column.}

\item{rdataclass}{Search "rdataclass" column.}

\item{sourcetype}{Search "sourcetype" column.}

\item{partial_match}{Return case-insensitive substring matches instead of 
exact matches only (default: \code{TRUE}).}

\item{peaks_only}{Return pre-filtered metadata for peak data 
aligned to the GRCh38 human genome build.}

\item{verbose}{Print messages.}

\item{...}{Additional metdata columns to filter.}
}
\value{
\link[data.table]{data.table}.
}
\description{
Search metadata from the Encyclopedia of DNA Elements
 (\href{https://www.encodeproject.org/}{ENCODE}).
}
\examples{
meta <- search_roadmap(biosample_type="brain",
                       file_format_type="narrowPeak",
                       genome="hg19",
                       peaks_only=TRUE)
}
