% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_wider_dt.R
\name{unnest_wider_dt}
\alias{unnest_wider_dt}
\title{Unnest data.table wider}
\source{
\href{https://stackoverflow.com/a/51448422/13214824}{
StackOverflow}
}
\usage{
unnest_wider_dt(dt, col, id_col, new_cols = NULL, verbose = TRUE)
}
\value{
\link[data.table]{data.table}.
}
\description{
A far more efficient version of\link[tidyr]{unnest_wider} for 
\link[data.table]{data.table}.
}
\keyword{internal}
