% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to.R
\name{bam_to}
\alias{bam_to}
\title{Convert a BAM file}
\usage{
bam_to(
  bam_files,
  formats = c("bedGraph", "bigWig"),
  width = 150,
  outdir = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{outdir}{Directory to save results to.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[rtracklayer]{export}.}
}
\value{
Converted file paths.
}
\description{
Convert a BAM file to any format supported by \link[rtracklayer]{export}.
}
\examples{
bam_files <- c(f1=system.file("extdata","querybins.bam",package="Rsamtools"),
               f2=system.file("extdata","revbins.bam",package="Rsamtools"))
converted_files <- bam_to(bam_files = bam_files)               
}
