% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to.R
\name{bed_to}
\alias{bed_to}
\title{Convert a BED file}
\usage{
bed_to(
  files,
  build = "hg19",
  keep_chr = NULL,
  formats = c("bigwig", "bigbed"),
  save_dir = tempdir(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{files}{Paths to one or more BED files.}

\item{build}{Genome build.}

\item{keep_chr}{Which chromosomes to keep.}

\item{save_dir}{Path to save bigBed files.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[rtracklayer]{export}.}
}
\value{
Converted file paths.
}
\description{
Convert a BED file to any format supported by \link[rtracklayer]{export}.
}
\examples{
files <- system.file("tests","test.bed",package = "rtracklayer")
out <- bed_to(files=files)
}
