% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_granges.R
\name{condense_granges}
\alias{condense_granges}
\title{Condense \link[GenomicRanges]{GRanges}}
\usage{
condense_granges(gr, split_chromosomes = TRUE)
}
\arguments{
\item{gr}{A \link[GenomicRanges]{GRanges} object.}

\item{split_chromosomes}{Split into a named list by chromosome.}
}
\description{
Condense a \link[GenomicRanges]{GRanges} object
by taking the min/max position per chromosome.
This helps to reduce the total number of queries, 
which can cause memory allocation problems 
due to repeated calls to the underlying C libraries.
}
\examples{
grlist <- list("gr1"=GenomicRanges::GRanges("4:1-100"),
               "gr2"=GenomicRanges::GRanges("4:200-500"),
               "gr3"=GenomicRanges::GRanges("8:-10000"))
gr2 <- condense_granges(gr = gr)                
}
