% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trackhub.R
\name{create_trackhub}
\alias{create_trackhub}
\title{Create TrackHub}
\source{
\href{UCSC track metadata descriptions}{
https://genome-euro.ucsc.edu/goldenPath/help/trackDb/trackDbHub.html
}

\href{example trackhub}{
https://github.com/mhalushka/UCSC_miRNA_barchart
}

\href{https://github.com/lawremi/rtracklayer/issues/12}{
rtracklayer possible memory leakage
}
}
\usage{
create_trackhub(
  hub = "cutntag_benchmarking",
  path = getwd(),
  save_dir = file.path(path, "trackhub"),
  folders = c(""),
  domain = "https://webserver-schilder-ukdri.dsi.ic.ac.uk/cutntag_benchmarking",
  shortLabel = "CUT&Tag benchmarking",
  longLabel =
    "Data track associated with the CUT&Tag benchmarking manuscript: https://doi.org/10.1101/2022.03.30.486382",
  email = "",
  visibility = 2,
  genome = "hg19",
  descriptionUrl = paste(save_dir, "description.html", sep = "/"),
  filetypes = c(bigWig = ".bigwig$|.bw$"),
  as_list = TRUE,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to search for files within recursively.}

\item{domain}{URL prefix where the files will be hosted.}

\item{genome}{Genome build that files were aligned to.}

\item{filetypes}{File types to search for.}

\item{as_list}{Return each file as a nested list.}

\item{folder}{Subfolders within \code{path} to search,}

\item{save_path}{Path to to save \emph{trackDb} file to.}
}
\value{
Character or nested list.
}
\description{
Create TrackHub for visualization of genomic files 
in the UCSC Genome Browser.
}
\examples{
\dontrun{
path <-"/Volumes/bms20/projects/neurogenomics-lab/live/Projects/CUT_n_TAG/CUTnTag_analysis"
res <- create_trackhub(path=path, folders <- c("bigwig","bedgraph","peaks"))
}
}
