% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_bg_bw.R
\name{example_bg_bw}
\alias{example_bg_bw}
\title{Example bedGraph/bigWig files}
\usage{
example_bg_bw(gsm = "GSM4703766", ranges = "chr6:165169213-167169213")
}
\arguments{
\item{gsm}{GEO GSM id.
Matches will be case-insensitive.}
}
\value{
Named list of file paths.
}
\description{
Download examples bedGraph and bigWig files.
}
\examples{
files <- example_bg_bw()
}
