% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome.R
\name{get_genome}
\alias{get_genome}
\title{Get genome}
\usage{
get_genome(
  keep.chr = paste0("chr", c(seq_len(22), "X", "Y")),
  genome = "hg19",
  sort = TRUE,
  style = "UCSC",
  split_chromosomes = FALSE
)
}
\arguments{
\item{genome}{The genome object or genome identifier.}
}
\description{
Get a genome (or a subset of chromosomes from a genome) 
as a \link[GenomicRanges]{GRanges} object. This can be useful for 
querying whole-chromosomes at a time.
}
\examples{
gr <- get_genome(keep.chr=1:2)
}
