% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_peaks.R
\name{pooled_peaks}
\alias{pooled_peaks}
\title{Pooled peaks}
\usage{
pooled_peaks(
  bam_files,
  groups = names(bam_files),
  save_dir = tempdir(),
  method = c("MACSr", "SEACR"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bam_files}{One or more BAM files.}

\item{groups}{A character vector of the same length as \code{bam_files}
defining how to group files when calling consensus peaks.}

\item{method}{Method to call consensus peaks with:
\itemize{
\item{"MACSr:}{Call consensus peaks by merging groups of BAM files with 
\link[Rsamtools]{mergeBam} and then calling peaks with 
\link[MACSr]{callpeak}.
}
}}
}
\value{
\link[GenomicRanges]{GRangesList} object.
}
\description{
Pool groups of BAM files into a smaller number of BAM files and call peaks
on each of them.
}
\examples{
bam_files <- example_bam()
peaks <- pooled_peaks(bam_files = bam_files)
}
